Finishing up the Search Screen UI

+++ app/src/main/java/com/devhp/firstcompose/model/VolumeInfo.kt
@@ -9,7 +9,7 @@ data class VolumeInfo(
     @SerializedName("authors")
     val authors: List<String?>? = listOf(),
     @SerializedName("averageRating")
-    val averageRating: Int? = 0,
+    val averageRating: Double? = 0.0,
     @SerializedName("canonicalVolumeLink")
     val canonicalVolumeLink: String? = "",
     @SerializedName("categories")
